Add knip unused checks
