+++ package.json
@@ -11,11 +11,12 @@
     "package": "plasmo package && node scripts/sync-locales.mjs",
     "lint": "oxlint --fix --deny-warnings src vitest.config.ts tests",
     "lint:check": "oxlint --deny-warnings src vitest.config.ts tests",
+    "knip": "knip",
     "format": "oxfmt --write .",
     "format:check": "oxfmt --check .",
     "test": "vitest",
     "test:ci": "vitest run",
-    "check": "pnpm format:check && pnpm lint:check && pnpm test:ci && pnpm build",
+    "check": "pnpm format:check && pnpm lint:check && pnpm knip && pnpm test:ci && pnpm build",
     "prepare": "lefthook install --reset-hooks-path",
     "e2e:run": "playwright test",
     "e2e:ui": "playwright test --ui"
@@ -24,7 +25,6 @@
     "@plasmohq/messaging": "^0.7.2",
     "@plasmohq/storage": "^1.15.0",
     "@radix-ui/react-slot": "^1.2.4",
-    "approve-builds": "^1.0.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^1.16.0",
@@ -35,7 +35,6 @@
     "tailwind-merge": "^3.6.0"
   },
   "devDependencies": {
-    "@iconify/json": "2.2.473",
     "@ngneat/falso": "8.0.2",
     "@playwright/test": "^1.60.0",
     "@types/chrome": "0.1.42",
@@ -45,6 +44,7 @@
     "@types/webextension-polyfill": "^0.12.5",
     "autoprefixer": "^10.5.0",
     "fuse.js": "7.3.0",
+    "knip": "^6.14.0",
     "lefthook": "2.1.6",
     "lint-staged": "17.0.4",
     "oxfmt": "0.50.0",
